[Feat] #18 보험 보장 항목을 abstract로 두고 @MappedSuperclass을 사용하여 상속받아 사용하도록 함.
