+++ src/main/java/org/sopt/bofit/domain/insurance/entity/benefit/Death.java
@@ -1,4 +1,4 @@
-package org.sopt.bofit.domain.insurance.entity;
+package org.sopt.bofit.domain.insurance.entity.benefit;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
@@ -6,13 +6,13 @@ import lombok.Getter;
 
 @Getter
 @Embeddable
-public class DailyHospitalization {
+public class Death {
 
-	@Column(name = "daily_hospitalization_disease", nullable = false,
+	@Column(name = "disease_death_benefit", nullable = false,
 		columnDefinition = "SMALLINT UNSIGNED DEFAULT 0")
 	private int disease;
 
-	@Column(name = "daily_hospitalization_injury", nullable = false,
+	@Column(name = "injury_death_benefit", nullable = false,
 		columnDefinition = "SMALLINT UNSIGNED DEFAULT 0")
 	private int injury;
 }
